# module: eliminate performance cost of ESM syntax detection for CommonJS entry points