--- a/lib/internal/modules/cjs/loader.js
+++ b/lib/internal/modules/cjs/loader.js
@@ -70,10 +70,18 @@ const cjsSourceCache = new SafeWeakMap();
  */
 const cjsExportsCache = new SafeWeakMap();
 
+/**
+ * Map of CJS modules where the initial attempt to parse as CommonJS failed;
+ * we want to retry as ESM but avoid reading the module from disk again.
+ * @type {SafeMap<string, string>} filename: contents
+ */
+const cjsRetryAsESMCache = new SafeMap();
+
 // Set first due to cycle with ESM loader functions.
 module.exports = {
   cjsExportsCache,
   cjsSourceCache,
+  cjsRetryAsESMCache,
   initializeCJS,
   Module,
   wrapSafe,
@@ -1337,6 +1345,11 @@ function wrapSafe(filename, content, cjsModuleInstance, codeCache) {
     return result;
   } catch (err) {
     if (process.mainModule === cjsModuleInstance) {
+      if (getOptionValue('--experimental-detect-module')) {
+        // For the main entry point, cache the source to potentially retry as ESM.
+        cjsRetryAsESMCache.set(filename, content);
+      }
+
       const { enrichCJSError } = require('internal/modules/esm/translators');
       enrichCJSError(err, content, filename);
     }