--- a/lib/internal/modules/cjs/loader.js
+++ b/lib/internal/modules/cjs/loader.js
@@ -1348,10 +1348,12 @@ function wrapSafe(filename, content, cjsModuleInstance, codeCache) {
       if (getOptionValue('--experimental-detect-module')) {
         // For the main entry point, cache the source to potentially retry as ESM.
         cjsRetryAsESMCache.set(filename, content);
+      } else {
+        // We only enrich the error (print a warning) if we're sure we're going to for-sure throw it; so if we're
+        // retrying as ESM, wait until we know whether we're going to retry before calling `enrichCJSError`.
+        const { enrichCJSError } = require('internal/modules/esm/translators');
+        enrichCJSError(err, content, filename);
       }
-
-      const { enrichCJSError } = require('internal/modules/esm/translators');
-      enrichCJSError(err, content, filename);
     }
     throw err;
   }