--- a/lib/internal/modules/cjs/loader.js
+++ b/lib/internal/modules/cjs/loader.js
@@ -70,18 +70,10 @@ const cjsSourceCache = new SafeWeakMap();
  */
 const cjsExportsCache = new SafeWeakMap();
 
-/**
- * Map of CJS modules where the initial attempt to parse as CommonJS failed;
- * we want to retry as ESM but avoid reading the module from disk again.
- * @type {SafeMap<string, string>} filename: contents
- */
-const cjsRetryAsESMCache = new SafeMap();
-
 // Set first due to cycle with ESM loader functions.
 module.exports = {
   cjsExportsCache,
   cjsSourceCache,
-  cjsRetryAsESMCache,
   initializeCJS,
   Module,
   wrapSafe,
@@ -1347,7 +1339,7 @@ function wrapSafe(filename, content, cjsModuleInstance, codeCache) {
     if (process.mainModule === cjsModuleInstance) {
       if (getOptionValue('--experimental-detect-module')) {
         // For the main entry point, cache the source to potentially retry as ESM.
-        cjsRetryAsESMCache.set(filename, content);
+        process.mainModule._retryAsESMSource = content;
       } else {
         // We only enrich the error (print a warning) if we're sure we're going to for-sure throw it; so if we're
         // retrying as ESM, wait until we know whether we're going to retry before calling `enrichCJSError`.