--- a/lib/internal/modules/cjs/loader.js
+++ b/lib/internal/modules/cjs/loader.js
@@ -70,14 +70,12 @@ const cjsSourceCache = new SafeWeakMap();
  */
 const cjsExportsCache = new SafeWeakMap();
 
-const kEntryPointSource = Symbol('kEntryPointSource');
-
 // Set first due to cycle with ESM loader functions.
 module.exports = {
   cjsExportsCache,
   cjsSourceCache,
   initializeCJS,
-  kEntryPointSource,
+  entryPointSource: undefined, // Set below.
   Module,
   wrapSafe,
 };
@@ -1342,7 +1340,7 @@ function wrapSafe(filename, content, cjsModuleInstance, codeCache) {
     if (process.mainModule === cjsModuleInstance) {
       if (getOptionValue('--experimental-detect-module')) {
         // For the main entry point, cache the source to potentially retry as ESM.
-        process.mainModule[kEntryPointSource] = content;
+        module.exports.entryPointSource = content;
       } else {
         // We only enrich the error (print a warning) if we're sure we're going to for-sure throw it; so if we're
         // retrying as ESM, wait until we know whether we're going to retry before calling `enrichCJSError`.