--- a/lib/internal/modules/helpers.js
+++ b/lib/internal/modules/helpers.js
@@ -320,6 +320,37 @@ function normalizeReferrerURL(referrerName) {
 }
 
 
+const esmSyntaxErrorMessages = new SafeSet([
+  'Cannot use import statement outside a module', // `import` statements
+  "Unexpected token 'export'",                    // `export` statements
+  "Cannot use 'import.meta' outside a module",    // `import.meta` references
+]);
+const throwsOnlyInCommonJSErrorMessages = new SafeSet([
+  "Identifier 'module' has already been declared",
+  "Identifier 'exports' has already been declared",
+  "Identifier 'require' has already been declared",
+  "Identifier '__filename' has already been declared",
+  "Identifier '__dirname' has already been declared",
+  'await is only valid in async functions and the top level bodies of modules', // Top-level `await`
+]);
+/**
+ * After an attempt to parse a module as CommonJS throws an error, should we try again as ESM?
+ * @param {string} errorMessage The string message thrown by V8 when attempting to parse as CommonJS
+ * @param {string} source Module contents
+ */
+function shouldRetryAsESM(errorMessage, source) {
+  if (esmSyntaxErrorMessages.has(errorMessage)) {
+    return true;
+  }
+
+  if (throwsOnlyInCommonJSErrorMessages.has(errorMessage)) {
+    return /** @type {boolean} */(internalBinding('contextify').shouldRetryAsESM(source));
+  }
+
+  return false;
+}
+
+
 // Whether we have started executing any user-provided CJS code.
 // This is set right before we call the wrapped CJS code (not after,
 // in case we are half-way in the execution when internals check this).
@@ -339,6 +370,7 @@ module.exports = {
   loadBuiltinModule,
   makeRequireFunction,
   normalizeReferrerURL,
+  shouldRetryAsESM,
   stripBOM,
   toRealPath,
   hasStartedUserCJSExecution() {