--- a/lib/internal/modules/helpers.js
+++ b/lib/internal/modules/helpers.js
@@ -19,6 +19,15 @@ const {
 } = require('internal/errors').codes;
 const { BuiltinModule } = require('internal/bootstrap/realm');
 
+const {
+  shouldRetryAsESM: contextifyShouldRetryAsESM,
+  constants: {
+    syntaxDetectionErrors: {
+      esmSyntaxErrorMessages,
+      throwsOnlyInCommonJSErrorMessages,
+    },
+  },
+} = internalBinding('contextify');
 const { validateString } = require('internal/validators');
 const fs = require('fs'); // Import all of `fs` so that it can be monkey-patched.
 const internalFS = require('internal/fs/utils');
@@ -320,31 +329,22 @@ function normalizeReferrerURL(referrerName) {
 }
 
 
-const esmSyntaxErrorMessages = new SafeSet([
-  'Cannot use import statement outside a module', // `import` statements
-  "Unexpected token 'export'",                    // `export` statements
-  "Cannot use 'import.meta' outside a module",    // `import.meta` references
-]);
-const throwsOnlyInCommonJSErrorMessages = new SafeSet([
-  "Identifier 'module' has already been declared",
-  "Identifier 'exports' has already been declared",
-  "Identifier 'require' has already been declared",
-  "Identifier '__filename' has already been declared",
-  "Identifier '__dirname' has already been declared",
-  'await is only valid in async functions and the top level bodies of modules', // Top-level `await`
-]);
+let esmSyntaxErrorMessagesSet; // Declared lazily in shouldRetryAsESM
+let throwsOnlyInCommonJSErrorMessagesSet; // Declared lazily in shouldRetryAsESM
 /**
  * After an attempt to parse a module as CommonJS throws an error, should we try again as ESM?
  * @param {string} errorMessage The string message thrown by V8 when attempting to parse as CommonJS
  * @param {string} source Module contents
  */
 function shouldRetryAsESM(errorMessage, source) {
-  if (esmSyntaxErrorMessages.has(errorMessage)) {
+  esmSyntaxErrorMessagesSet ??= new SafeSet(esmSyntaxErrorMessages);
+  if (esmSyntaxErrorMessagesSet.has(errorMessage)) {
     return true;
   }
 
-  if (throwsOnlyInCommonJSErrorMessages.has(errorMessage)) {
-    return /** @type {boolean} */(internalBinding('contextify').shouldRetryAsESM(source));
+  throwsOnlyInCommonJSErrorMessagesSet ??= new SafeSet(throwsOnlyInCommonJSErrorMessages);
+  if (throwsOnlyInCommonJSErrorMessagesSet.has(errorMessage)) {
+    return /** @type {boolean} */(contextifyShouldRetryAsESM(source));
   }
 
   return false;