--- a/lib/internal/modules/helpers.js
+++ b/lib/internal/modules/helpers.js
@@ -333,6 +333,9 @@ let esmSyntaxErrorMessagesSet; // Declared lazily in shouldRetryAsESM
 let throwsOnlyInCommonJSErrorMessagesSet; // Declared lazily in shouldRetryAsESM
 /**
  * After an attempt to parse a module as CommonJS throws an error, should we try again as ESM?
+ * We only want to try again as ESM if the error is due to syntax that is only valid in ESM; and if the CommonJS parse
+ * throws on an error that would not have been a syntax error in ESM (like via top-level `await` or a lexical
+ * redeclaration of one of the CommonJS variables) then we need to parse again to see if it would have thrown in ESM.
  * @param {string} errorMessage The string message thrown by V8 when attempting to parse as CommonJS
  * @param {string} source Module contents
  */