--- a/lib/internal/modules/run_main.js
+++ b/lib/internal/modules/run_main.js
@@ -173,6 +173,10 @@ function executeUserEntryPoint(main = process.argv[1]) {
         cjsRetryAsESMCache.delete(resolvedMain);
         const { shouldRetryAsESM } = require('internal/modules/helpers');
         retryAsESM = shouldRetryAsESM(error.message, source);
+        if (!retryAsESM) {
+          const { enrichCJSError } = require('internal/modules/esm/translators');
+          enrichCJSError(error, source, resolvedMain);
+        }
       }
       if (!retryAsESM) {
         throw error;