--- a/lib/internal/modules/run_main.js
+++ b/lib/internal/modules/run_main.js
@@ -164,10 +164,10 @@ function executeUserEntryPoint(main = process.argv[1]) {
   if (!useESMLoader) {
     // Module._load is the monkey-patchable CJS module loader.
     const { Module } = require('internal/modules/cjs/loader');
-    try {
-      Module._load(main, null, true);
-    } catch (error) {
-      if (getOptionValue('--experimental-detect-module')) {
+    if (getOptionValue('--experimental-detect-module')) {
+      try {
+        Module._load(main, null, true);
+      } catch (error) {
         const { cjsRetryAsESMCache } = require('internal/modules/cjs/loader');
         const source = cjsRetryAsESMCache.get(resolvedMain);
         cjsRetryAsESMCache.delete(resolvedMain);
@@ -176,11 +176,11 @@ function executeUserEntryPoint(main = process.argv[1]) {
         if (!retryAsESM) {
           const { enrichCJSError } = require('internal/modules/esm/translators');
           enrichCJSError(error, source, resolvedMain);
+          throw error;
         }
       }
-      if (!retryAsESM) {
-        throw error;
-      }
+    } else {
+      Module._load(main, null, true);
     }
   }