--- a/lib/internal/modules/run_main.js
+++ b/lib/internal/modules/run_main.js
@@ -5,7 +5,6 @@ const {
   globalThis,
 } = primordials;
 
-const { containsModuleSyntax } = internalBinding('contextify');
 const { getNearestParentPackageJSONType } = internalBinding('modules');
 const { getOptionValue } = require('internal/options');
 const { checkPackageJSONIntegrity } = require('internal/modules/package_json_reader');
@@ -87,10 +86,6 @@ function shouldUseESMLoader(mainPath) {
 
   // No package.json or no `type` field.
   if (response === undefined || response[0] === 'none') {
-    if (getOptionValue('--experimental-detect-module')) {
-      // If the first argument of `containsModuleSyntax` is undefined, it will read `mainPath` from the file system.
-      return containsModuleSyntax(undefined, mainPath);
-    }
     return false;
   }
 
@@ -162,7 +157,30 @@ function runEntryPointWithESMLoader(callback) {
 function executeUserEntryPoint(main = process.argv[1]) {
   const resolvedMain = resolveMainPath(main);
   const useESMLoader = shouldUseESMLoader(resolvedMain);
-  if (useESMLoader) {
+
+  // We might retry as ESM if the CommonJS loader throws because the file is ESM, so
+  // try the CommonJS loader first unless we know it's supposed to be parsed as ESM.
+  let retryAsESM = false;
+  if (!useESMLoader) {
+    // Module._load is the monkey-patchable CJS module loader.
+    const { Module } = require('internal/modules/cjs/loader');
+    try {
+      Module._load(main, null, true);
+    } catch (error) {
+      if (getOptionValue('--experimental-detect-module')) {
+        const { cjsRetryAsESMCache } = require('internal/modules/cjs/loader');
+        const source = cjsRetryAsESMCache.get(resolvedMain);
+        cjsRetryAsESMCache.delete(resolvedMain);
+        const { shouldRetryAsESM } = require('internal/modules/helpers');
+        retryAsESM = shouldRetryAsESM(error.message, source);
+      }
+      if (!retryAsESM) {
+        throw error;
+      }
+    }
+  }
+
+  if (useESMLoader || retryAsESM) {
     const mainPath = resolvedMain || main;
     const mainURL = pathToFileURL(mainPath).href;
 
@@ -171,10 +189,6 @@ function executeUserEntryPoint(main = process.argv[1]) {
       // even after the event loop stops running.
       return cascadedLoader.import(mainURL, undefined, { __proto__: null }, true);
     });
-  } else {
-    // Module._load is the monkey-patchable CJS module loader.
-    const { Module } = require('internal/modules/cjs/loader');
-    Module._load(main, null, true);
   }
 }