--- a/lib/internal/modules/run_main.js
+++ b/lib/internal/modules/run_main.js
@@ -168,9 +168,9 @@ function executeUserEntryPoint(main = process.argv[1]) {
       try {
         Module._load(main, null, true);
       } catch (error) {
-        const { cjsRetryAsESMCache } = require('internal/modules/cjs/loader');
-        const source = cjsRetryAsESMCache.get(resolvedMain);
-        cjsRetryAsESMCache.delete(resolvedMain);
+        const source = process.mainModule._retryAsESMSource;
+        // In case the entry point is a large file, such as a bundle, release it from memory since we're done with it.
+        process.mainModule._retryAsESMSource = undefined; // No need to `delete` since we don't care if the key exists.
         const { shouldRetryAsESM } = require('internal/modules/helpers');
         retryAsESM = shouldRetryAsESM(error.message, source);
         if (!retryAsESM) {