--- a/lib/internal/modules/run_main.js
+++ b/lib/internal/modules/run_main.js
@@ -152,14 +152,16 @@ function runEntryPointWithESMLoader(callback) {
  * by `require('module')`) even when the entry point is ESM.
  * This monkey-patchable code is bypassed under `--experimental-default-type=module`.
  * Because of backwards compatibility, this function is exposed publicly via `import { runMain } from 'node:module'`.
+ * When `--experimental-detect-module` is passed, this function will attempt to run ambiguous (no explicit extension, no
+ * `package.json` type field) entry points as CommonJS first; under certain conditions, it will retry running as ESM.
  * @param {string} main - First positional CLI argument, such as `'entry.js'` from `node entry.js`
  */
 function executeUserEntryPoint(main = process.argv[1]) {
   const resolvedMain = resolveMainPath(main);
   const useESMLoader = shouldUseESMLoader(resolvedMain);
 
-  // We might retry as ESM if the CommonJS loader throws because the file is ESM, so
-  // try the CommonJS loader first unless we know it's supposed to be parsed as ESM.
+  // Unless we know we should use the ESM loader to handle the entry point per the checks in `shouldUseESMLoader`, first
+  // try to run the entry point via the CommonJS loader; and if that fails under certain conditions, retry as ESM.
   let retryAsESM = false;
   if (!useESMLoader) {
     // Module._load is the monkey-patchable CJS module loader.
@@ -168,9 +170,10 @@ function executeUserEntryPoint(main = process.argv[1]) {
       try {
         Module._load(main, null, true);
       } catch (error) {
-        const source = process.mainModule._retryAsESMSource;
+        const { kEntryPointSource } = require('internal/modules/cjs/loader');
+        const source = process.mainModule[kEntryPointSource];
         // In case the entry point is a large file, such as a bundle, release it from memory since we're done with it.
-        process.mainModule._retryAsESMSource = undefined; // No need to `delete` since we don't care if the key exists.
+        process.mainModule[kEntryPointSource] = undefined;
         const { shouldRetryAsESM } = require('internal/modules/helpers');
         retryAsESM = shouldRetryAsESM(error.message, source);
         if (!retryAsESM) {
@@ -179,7 +182,7 @@ function executeUserEntryPoint(main = process.argv[1]) {
           throw error;
         }
       }
-    } else {
+    } else { // `--experimental-detect-module` is not passed
       Module._load(main, null, true);
     }
   }