--- a/lib/internal/modules/run_main.js
+++ b/lib/internal/modules/run_main.js
@@ -164,18 +164,18 @@ function executeUserEntryPoint(main = process.argv[1]) {
   // try to run the entry point via the CommonJS loader; and if that fails under certain conditions, retry as ESM.
   let retryAsESM = false;
   if (!useESMLoader) {
-    // Module._load is the monkey-patchable CJS module loader.
-    const { Module } = require('internal/modules/cjs/loader');
+    const cjsLoader = require('internal/modules/cjs/loader');
+    const { Module } = cjsLoader;
     if (getOptionValue('--experimental-detect-module')) {
       try {
+        // Module._load is the monkey-patchable CJS module loader.
         Module._load(main, null, true);
       } catch (error) {
-        const { kEntryPointSource } = require('internal/modules/cjs/loader');
-        const source = process.mainModule[kEntryPointSource];
-        // In case the entry point is a large file, such as a bundle, release it from memory since we're done with it.
-        process.mainModule[kEntryPointSource] = undefined;
+        const source = cjsLoader.entryPointSource;
         const { shouldRetryAsESM } = require('internal/modules/helpers');
         retryAsESM = shouldRetryAsESM(error.message, source);
+        // In case the entry point is a large file, such as a bundle, release it from memory since we're done with it.
+        cjsLoader.entryPointSource = undefined;
         if (!retryAsESM) {
           const { enrichCJSError } = require('internal/modules/esm/translators');
           enrichCJSError(error, source, resolvedMain);