--- a/lib/internal/modules/run_main.js
+++ b/lib/internal/modules/run_main.js
@@ -174,7 +174,8 @@ function executeUserEntryPoint(main = process.argv[1]) {
         const source = cjsLoader.entryPointSource;
         const { shouldRetryAsESM } = require('internal/modules/helpers');
         retryAsESM = shouldRetryAsESM(error.message, source);
-        // In case the entry point is a large file, such as a bundle, release it from memory since we're done with it.
+        // In case the entry point is a large file, such as a bundle,
+        // ensure no further references can prevent it being garbage-collected.
         cjsLoader.entryPointSource = undefined;
         if (!retryAsESM) {
           const { enrichCJSError } = require('internal/modules/esm/translators');