--- a/src/node_contextify.h
+++ b/src/node_contextify.h
@@ -106,6 +106,7 @@ class ContextifyContext : public BaseObject {
       const v8::ScriptCompiler::Source& source);
   static void ContainsModuleSyntax(
       const v8::FunctionCallbackInfo<v8::Value>& args);
+  static void ShouldRetryAsESM(const v8::FunctionCallbackInfo<v8::Value>& args);
   static void WeakCallback(
       const v8::WeakCallbackInfo<ContextifyContext>& data);
   static void PropertyGetterCallback(