… benchmark/misc/startup-core.js
… test/es-module/test-esm-detect-ambiguous.mjs