@joyeecheung 2024-03-15
Can you add the typescript.js bundle into the startup-core or startup-cli-version benchmark and run it in the benchmark CI?