@joyeecheung 2024-03-15
> Can we flip the flag in v22?

I think we should only flip the flag if we print a warning when the fallback is used. Users should mark the ESM explicitly wherever possible and be aware of the performance cost of the fallback. You probably don’t want thousands of your implicit .js ESMs to all get parsed as CJS first before falling back to ESM silently and wonder why your module loading is so slow…