@mcollina 2024-03-15
Agreed on the warning. It would be great to ship it given it only simplifies dx.