@GeoffreyBooth 2024-03-15
Regarding the warning, I think we definitely want to _avoid_ warning for two scenarios:

- An extensionless file outside of any `package.json` scope, like `/usr/local/bin/foo`
- Any files within `node_modules`

The user can’t add explicitness for the former without breaking their desired UX, and the user can’t fix the latter without patching dependencies.

The case I’m unsure on is `node entry.js` where there’s no `package.json` file (but there is an extension). The user _could_ change the .js to .mjs or .cjs, but I’m not sure we want to push them to do so; they may have great reasons for needing to use .js (their editor, other tools, etc.).

How about we warn only on ambiguous entry points where there _is_ a `package.json` present? That probably covers most usage of detection, and if the user already has a `package.json` there shouldn’t be a reason that they can’t add a `type` field to it (whether that field’s value is `module` or `commonjs`).