@targos 2024-03-18
@GeoffreyBooth Here's what I was able to write to print the original error location:

<details>

```diff
diff --git a/lib/internal/modules/run_main.js b/lib/internal/modules/run_main.js
index 8c29051034..b9f6bba3b0 100644
--- a/lib/internal/modules/run_main.js
+++ b/lib/internal/modules/run_main.js
@@ -16,6 +16,7 @@ const {
 } = require('internal/process/execution');
 const {
   triggerUncaughtException,
+  tryCatchRethrow,
 } = internalBinding('errors');
 const {
   privateSymbols: {
@@ -164,9 +165,9 @@ function executeUserEntryPoint(main = process.argv[1]) {
   if (!useESMLoader) {
     // Module._load is the monkey-patchable CJS module loader.
     const { Module } = require('internal/modules/cjs/loader');
-    try {
+    tryCatchRethrow(() => {
       Module._load(main, null, true);
-    } catch (error) {
+    }, (error) => {
       if (getOptionValue('--experimental-detect-module')) {
         const { cjsRetryAsESMCache } = require('internal/modules/cjs/loader');
         const source = cjsRetryAsESMCache.get(resolvedMain);
@@ -178,10 +179,8 @@ function executeUserEntryPoint(main = process.argv[1]) {
           enrichCJSError(error, source, resolvedMain);
         }
       }
-      if (!retryAsESM) {
-        throw error;
-      }
-    }
+      return !retryAsESM;
+    });
   }
 
   if (useESMLoader || retryAsESM) {
diff --git a/src/node_errors.cc b/src/node_errors.cc
index ff091fd20d..bfeed155d5 100644
--- a/src/node_errors.cc
+++ b/src/node_errors.cc
@@ -1088,6 +1088,39 @@ static void TriggerUncaughtException(const FunctionCallbackInfo<Value>& args) {
   errors::TriggerUncaughtException(isolate, exception, message, from_promise);
 }
 
+// Takes two JS callbacks, and calls the first one in a TryCatch scope.
+// If the first one throws, call the second one with the exception.
+// If the second one returns true, rethrow the original exception.
+// This is to keep the location of the original throw in JS.
+static void TryCatchRethrow(const FunctionCallbackInfo<Value>& args) {
+  CHECK(args.Length() == 2);
+  CHECK(args[0]->IsFunction());
+  CHECK(args[1]->IsFunction());
+
+  Local<Function> try_fn = args[0].As<Function>();
+  Local<Function> catch_fn = args[1].As<Function>();
+
+  Isolate* isolate = args.GetIsolate();
+  Environment* env = Environment::GetCurrent(isolate);
+  Local<Context> context = isolate->GetCurrentContext();
+
+  TryCatchScope try_catch_scope(env);
+  USE(try_fn->Call(context, Undefined(isolate), 0, nullptr));
+
+  if (try_catch_scope.HasCaught()) {
+    Local<Value> exception = try_catch_scope.Exception();
+    Local<Value> argv[] = {exception};
+    Local<Value> result;
+    if (!catch_fn->Call(context, Undefined(isolate), 1, argv)
+             .ToLocal(&result)) {
+      return;
+    }
+    if (result->IsTrue()) {
+      try_catch_scope.ReThrow();
+    }
+  }
+}
+
 void RegisterExternalReferences(ExternalReferenceRegistry* registry) {
   registry->Register(SetPrepareStackTraceCallback);
   registry->Register(SetGetSourceMapErrorSource);
@@ -1096,6 +1129,7 @@ void RegisterExternalReferences(ExternalReferenceRegistry* registry) {
   registry->Register(SetEnhanceStackForFatalException);
   registry->Register(NoSideEffectsToString);
   registry->Register(TriggerUncaughtException);
+  registry->Register(TryCatchRethrow);
 }
 
 void Initialize(Local<Object> target,
@@ -1123,6 +1157,7 @@ void Initialize(Local<Object> target,
       context, target, "noSideEffectsToString", NoSideEffectsToString);
   SetMethod(
       context, target, "triggerUncaughtException", TriggerUncaughtException);
+  SetMethod(context, target, "tryCatchRethrow", TryCatchRethrow);
 
   Isolate* isolate = context->GetIsolate();
   Local<Object> exit_codes = Object::New(isolate);
```

</details>