@joyeecheung 2025-02-06
Backport in https://github.com/nodejs/node/pull/56927